pyproject.toml
[tool.poetry]
name = "reflex_counter"
version = "1.0.9"
description = ""
authors = ["Ernad Husremovic <hernad@bring.out.ba>"]
readme = "README.md"

[tool.poetry.scripts]
reflex_counter = "reflex_counter.main:start"

[tool.poetry.dependencies]
python = "^3.12"
reflex = "*"
fastapi = "*"
python-keycloak = "^4.7.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
